Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy_utils
psycopg2-binary
python-dotenv
pandas
google-generativeai
python-multipart
pydantic
requests
